+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 * Remove `inline` attributes on trait method side. It can avoid [`clippy::inline_fn_without_body`](https://rust-lang.github.io/rust-clippy/master/index.html#inline_fn_without_body) by this.
 
+* Use `#[allow(patterns_in_fns_without_body)]` to generated extension trait.
+
+* Fix some bugs related to generics.
+
 # 0.1.1 - 2019-02-21
 
 * Fix an error related to generics.