+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Unreleased
 
-# 0.1.2 - 2019-02-21
-
 * Remove `inline` attributes on trait method side. It can avoid [`clippy::inline_fn_without_body`](https://rust-lang.github.io/rust-clippy/master/index.html#inline_fn_without_body) by this.
+
+# 0.1.2 - 2019-02-21
 
 * Use `#[allow(patterns_in_fns_without_body)]` to generated extension trait.